Update references module

+++ Controllers/AnexoReferenciaController.cs
@@ -19,45 +19,85 @@ namespace RepositorioDocumentos.Controllers
                 if (Session["userID"] == null) throw new Exception("505: Por favor intente logearse de nuevo en el sistema. (La Sesión expiró)");
 
                 HttpPostedFileBase file = Request.Files["file"];
+                string description = Request.Form["description"];
+                string filePath = Request.Form["filePath"];
+                int documentHeaderId = int.Parse(Request.Form["DocumentHeaderId"]);
+                int userId = int.Parse(Session["userID"].ToString());
+                string referenceType = Request.Form["referenceType"];
 
-                if (file != null && file.ContentLength > 0)
+                if (referenceType == "Link")
                 {
-                    string description = Request.Form["description"];
-                    int documentHeaderId = int.Parse(Request.Form["DocumentHeaderId"]);
-                    int userId = int.Parse(Session["userID"].ToString());
-
-                    string fileName = Path.GetFileName(file.FileName);
-                    string newFileName = $"{Path.GetFileNameWithoutExtension(file.FileName)}_{documentHeaderId}{Path.GetExtension(fileName)}";
-                    string filePath = Path.Combine(Server.MapPath("~/AnexosReferencias"), newFileName);
-
-                    file.SaveAs(filePath);
-
                     try
                     {
                         using (var db = new RepositorioDocRCEntities())
                         {
+                            var referenceExist = db.DocumentReferences.Where(r => r.DocumentHeaderId == documentHeaderId 
+                                                                             && r.Name.ToLower() == description.ToLower()).FirstOrDefault();
+                            if (referenceExist != null)
+                                return Json(new { result = "500", message = "Existe un anexo con esta descripción/título." });
+
                             db.DocumentReferences.Add(new DocumentReference
                             {
                                 DocumentHeaderId = documentHeaderId,
+                                ReferenceType = referenceType,
                                 Name = description,
-                                Url = newFileName,
+                                Url = filePath,
                                 CreatedDate = DateTime.Now,
                                 CreatedBy = userId
                             });
                             db.SaveChanges();
                         }
+
+                        return Json(new { result = "200", message = "success" });
                     }
                     catch (Exception ex)
                     {
                         Helper.SendException(ex);
                     }
-
-                    return Json(new { result = "200", message = "success" });
-                }
+                } 
                 else
                 {
-                    return Json(new { result = "404", message = "No ha seleccionado" });
-                }
+                    if (file != null && file.ContentLength > 0)
+                    {
+                        try
+                        {
+                            using (var db = new RepositorioDocRCEntities())
+                            {
+                                var referenceExist = db.DocumentReferences.Where(r => r.DocumentHeaderId == documentHeaderId
+                                                                         && r.Name.ToLower() == description.ToLower()).FirstOrDefault();
+                                if (referenceExist != null)
+                                    return Json(new { result = "500", message = "Existe un anexo con esta descripción/título." });
+
+                                string fileName = Path.GetFileName(file.FileName);
+                                string newFileName = $"{Path.GetFileNameWithoutExtension(file.FileName)}_{documentHeaderId}{Path.GetExtension(fileName)}";
+                                string _filePath = Path.Combine(Server.MapPath("~/AnexosReferencias"), newFileName);
+
+                                file.SaveAs(_filePath);
+
+                                db.DocumentReferences.Add(new DocumentReference
+                                {
+                                    DocumentHeaderId = documentHeaderId,
+                                    ReferenceType= referenceType,
+                                    Name = description,
+                                    Url = newFileName,
+                                    CreatedDate = DateTime.Now,
+                                    CreatedBy = userId
+                                });
+                                db.SaveChanges();
+                            }
+                        }
+                        catch (Exception ex)
+                        {
+                            Helper.SendException(ex);
+                        }
+
+                        return Json(new { result = "200", message = "success" });
+                    }
+                    else
+                    {
+                        return Json(new { result = "404", message = "No ha seleccionado" });
+                    }
+                } 
             }
             catch (Exception ex)
             {
@@ -65,6 +105,8 @@ namespace RepositorioDocumentos.Controllers
 
                 return Json(new { result = "500", message = ex.Message });
             }
+
+            return Json(new { result = "500", message = "Favor revisar si los campos fueron completados correctamente." });
         }
 
         [HttpPost]
@@ -76,7 +118,7 @@ namespace RepositorioDocumentos.Controllers
                 {
                     var references = db.DocumentReferences
                                        .Where(d => d.DocumentHeaderId == documentHeaderId)
-                                       .Select(s => new {s.Id, s.Url, s.Name})
+                                       .Select(s => new {s.Id, s.ReferenceType, s.Url, s.Name})
                                        .ToList();
                     return Json(new { result = "200", message = references });
                 }
@@ -104,12 +146,15 @@ namespace RepositorioDocumentos.Controllers
                         db.DocumentReferences.Remove(reference);
                         db.SaveChanges();
 
-                        string filePathToDelete = Server.MapPath($"~/AnexosReferencias/{reference.Url}");
-                        if (System.IO.File.Exists(filePathToDelete))
+                        if (reference.ReferenceType == "Archivo")
                         {
-                            System.IO.File.Delete(filePathToDelete);
-                            return Json(new { result = "200", message = "success" });
+                            string filePathToDelete = Server.MapPath($"~/AnexosReferencias/{reference.Url}");
+                            if (System.IO.File.Exists(filePathToDelete))
+                            {
+                                System.IO.File.Delete(filePathToDelete);
+                            }
                         }
+                        return Json(new { result = "200", message = "success" });
                     }
                 }
             }
@@ -120,7 +165,7 @@ namespace RepositorioDocumentos.Controllers
                 return Json(new { result = "500", message = ex.Message });
             }
 
-            return Json(new { result = "404", message = "El archivo no fue encontrado." });
+            return Json(new { result = "404", message = "El anexo no fue encontrado." });
         }
     }
 }
